codeforces

+++ Starting_Dsa/BASICS.cpp/CodeForces/Upsolve/midoriya.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    int x,y;
+    long long x,y;
     cin>>x>>y;
     cout<<x*y;
     return 0;
